backend/requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-redis==5.4.0
django-storages==1.14.2

# Django Channels (WebSockets)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
firebase-admin==6.4.0
PyJWT==2.8.0
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5

# Task Queue
celery==5.3.4
redis==5.0.1

# File Storage
boto3==1.29.7

# AI/ML
openai==1.3.7
google-generativeai==0.8.3

# GitHub Integration
PyGithub==1.59.1
requests==2.31.0

# PDF Generation
reportlab==4.0.7
WeasyPrint==60.2
PyPDF2==3.0.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
coverage==7.3.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
python-dateutil==2.8.2

